Reorganised components in folders, renamed some files
