+++ client/src/components/event/EventListItemModal.js
