+++ client/src/components/searchBar/SearchBar.js
