+++ client/src/pages/CalendarPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import CalendarComponent from "../components/CalendarComponent";
+import CalendarComponent from "../components/calendar/CalendarComponent";
 
 const CalendarPage = () => {
   return <CalendarComponent />;
